# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Rukhsat API
  version: 1.0.0
  description: >
    Contract between the backend and all three frontends (parent app, staff
    app, admin web + classroom display).


    This file is the referee. `packages/shared/src/types/api.ts` and
    `backend/app/schemas.py` describe the same shapes; when they disagree with
    this file, they are the bug.


    Change deliberately and tell the other developer before implementing —
    this is what lets both people build in parallel without blocking.
servers:
  - url: https://api.tideover.site/v1
    description: Production
  - url: http://localhost:8000/v1
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [status, code, message]
      properties:
        status: { type: integer, example: 403 }
        code: { type: string, example: not_authorized_for_student }
        message: { type: string }
        details: { type: object, additionalProperties: true, nullable: true }
    Role:
      type: string
      description: >
        `driver` collects children from many families and is never a guardian
        of any of them.
      enum: [parent, teacher, guard, admin, driver]
    Locale:
      type: string
      enum: [en, ur]
...
    School:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        lat: { type: number }
        lng: { type: number }
        geofence_radius_m: { type: integer }
        dismissal_time: { type: string, example: "13:15" }
        timezone: { type: string, example: Asia/Karachi }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        school_id: { type: string, format: uuid }
        role: { $ref: '#/components/schemas/Role' }
        name: { type: string }
        name_ur: { type: string, nullable: true }
        phone: { type: string }
        locale: { $ref: '#/components/schemas/Locale' }
        photo_url: { type: string, nullable: true }
...
paths:
...
  /auth/register-admin:
    post:
      tags: [auth]
      security: []
      summary: Register a school and its first administrator, together
      description: >
        One request creates both, in one transaction. `users.school_id` is NOT
        NULL, so an administrator cannot exist before their school — rather
        than weaken a constraint the whole authorization model leans on, signup
        takes both halves at once and the UI collects them across two screens.


        The new school is given its own ES256 keypair here, not lazily: a
        school without one looks completely healthy until the first parent
        shows a code at the gate and it cannot be verified.


        Returns a token, so the browser goes straight to the dashboard instead
        of bouncing back to a login form.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, phone, password, school]
              properties:
                name: { type: string, minLength: 2 }
                name_ur: { type: string, nullable: true }
                phone: { type: string }
                password: { type: string, minLength: 8 }
                locale: { type: string, enum: [en, ur], default: en }
                school:
                  type: object
                  required: [name, lat, lng]
                  properties:
                    name: { type: string, minLength: 2 }
                    lat: { type: number, minimum: -90, maximum: 90 }
                    lng: { type: number, minimum: -180, maximum: 180 }
                    geofence_radius_m:
                      type: integer
                      minimum: 100
                      maximum: 20000
                      default: 1000
                      description: >
                        Bounded deliberately. Below ~100 m the ring sits inside
                        the building and nobody ever enters it; above 20 km it
                        covers a city and stops meaning "nearly here".
                    dismissal_time: { type: string, example: "13:15" }
                    timezone: { type: string, default: Asia/Karachi }
      responses:
        '201':
          description: School and administrator created
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  expires_in: { type: integer }
                  user: { $ref: '#/components/schemas/User' }
                  school: { $ref: '#/components/schemas/School' }
        '409': { description: Phone number already registered }
        '422': { description: Invalid geofence radius or dismissal time }
...
  /schools/{school_id}:
    patch:
      tags: [schools]
      summary: Move the school or resize its geofence
      description: >
        Administrators, and only for their own school. The geofence radius
        governs whether arrival announcements fire at all, so this is a much
        larger lever than the screen suggests. `lat` and `lng` must be sent
        together — half a coordinate pair puts the school in the sea and the
        geofence silently stops matching anyone.
      parameters:
        - { name: school_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 2 }
                lat: { type: number, minimum: -90, maximum: 90 }
                lng: { type: number, minimum: -180, maximum: 180 }
                geofence_radius_m: { type: integer, minimum: 100, maximum: 20000 }
                dismissal_time: { type: string, example: "13:15" }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/School' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...